# server/contracts/openapi/homeos.yaml
...
openapi: "3.0.3"
info:
  title: HomeOS API
  version: "1.0"
  description: >
    HomeOS 底座服务接口，负责身份、家庭、权限、通知、同步、搜索等基础能力。
    所有接口均需携带 JWT token（除健康检查与 JWKS 分发外）。
servers:
  - url: /api/homeos
paths:
...
  /notifications:
    get:
      summary: 通知列表
      description: 按 type 分组返回通知列表与未读计数
      operationId: getNotifications
      tags:
        - notifications
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [budget_alert, system, reminder]
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 通知列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        type:
                          type: string
                          enum: [budget_alert, system, reminder]
                        content:
                          type: string
                        read_at:
                          type: string
                          format: date-time
                          nullable: true
                        created_at:
                          type: string
                          format: date-time
                  unread:
                    type: integer
                  unread_by_type:
                    type: object
                    description: >
                      本次新增的附加字段（消息页三个 type 页签的计数，服务端已由同一 GROUP BY 算出）。
                      三个键恒在，无未读的类型为 0；其值与同级 unread 一样按当前会话成员统计，
                      不受 ?type= 筛选影响（unread 亦为全量总数，PRD 17.1 第 2 条口径唯一，
                      页签筛选不得改动顶栏/首页红点读数）。计数一律服务端计算，请求不接受客户端上传。
                    properties:
                      budget_alert:
                        type: integer
                      system:
                        type: integer
                      reminder:
                        type: integer
                  next_cursor:
                    type: string
                    nullable: true
  /notifications/read:
    post:
      summary: 标记通知已读
      description: 支持单个通知已读与全部已读
      operationId: markNotificationsRead
      tags:
        - notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notification_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: 为空则标记全部已读
                all:
                  type: boolean
                  default: false
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  marked_count:
                    type: integer
                  unread:
                    type: integer

  # === 全局搜索 ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT